wapp/requirements.txt
easyocr==1.7.1
Flask==3.0.3
Flask-RESTful==0.3.10
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.2
scikit-image==0.23.2
scipy==1.13.1
urllib3==2.2.1
